+++ src/cjs/Circle.js
@@ -18,7 +18,8 @@
  * @modified 2026-06-10 Adding the utility function `Circle.circleUtils.containsPoint`.
  * @modified 2026-06-10 Adding the `Circle.clone` method.
  * @modified 2026-01-13 Adding helper function `Circle.circleUtils.containsPoint` and refactored the member method `containsPoint`.
- * @version  1.6.0
+ * @modified 2026-07-03 Adding the optional `epsilon` parameter to the `Circle.containsCircle` method.
+ * @version  1.7.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Circle = void 0;
@@ -92,8 +93,13 @@ var Circle = /** @class */ (function () {
      * @memberof Circle
      * @return {boolean} `true` if any only if the given circle is completely inside this circle.
      */
-    Circle.prototype.containsCircle = function (circle) {
-        return this.center.distance(circle.center) + circle.radius < this.radius;
+    Circle.prototype.containsCircle = function (circle, epsilon) {
+        if (typeof epsilon === "undefined" || Number.isNaN(epsilon)) {
+            return this.center.distance(circle.center) + circle.radius < this.radius;
+        }
+        else {
+            return this.center.distance(circle.center) + circle.radius < this.radius + Math.abs(epsilon);
+        }
     };
     /**
      * Calculate the distance from this circle to the given line.