+++ changelog.md
@@ -8,6 +8,9 @@
 - Add 'Rulers' class from the ngdg project.
 - Add 'onclose' and 'onopen' to the Modal (demos).
 
+- 2026-07-03
+  - Adding `utils/algorithms/solveApollonius3` to the library.
+  - Adding the optional `epsilon` parameter to the `Circle.containsCircle` method.
 - 2026-06-13
   - Adding `Triangle.getVertices()`.
   - Adding helper function `Circle.circleUtils.containsPoint` and refactored the member method `containsPoint`.